fix : 엔티티 데이터타입 중 Long -> Double으로 수정 - 식당 위도경도 Double로 수정 - 리뷰 별점 Double로 수정

+++ src/main/java/com/wandted/matitnyam/restaurant/dto/RestaurantDetailResponse.java
@@ -7,7 +7,7 @@ public record RestaurantDetailResponse(
     String name,
     District district,
     String address,
-    Long latitude,
-    Long longitude
+    Double latitude,
+    Double longitude
 ) {
 }
